new logger implementation and NetworkInterceptor movement
